indoormap test
