+++ tests/IndoorMap.test.tsx
@@ -1,398 +1,286 @@
-/**
- * @jest-environment jsdom
- */
-
 import React from 'react';
-import { render, screen, act } from '@testing-library/react';
-import '@testing-library/jest-dom';
-import { CoordsProvider, useCoords } from '../src/data/CoordsContext';
-import { IndoorsProvider, useIndoor } from '../src/data/IndoorContext';
-import { BuildingFloorAssociation } from '../src/interfaces/buildingFloorAssociation';
+import { render } from '@testing-library/react';
+import { HighlightIndoorMap } from '../src/components/IndoorMap';
+import { useCoords } from '../src/data/CoordsContext';
+import { useIndoor } from '../src/data/IndoorContext';
+import Mapbox from '@rnmapbox/maps';
 
-// Test component for CoordsContext
-const TestCoordsComponent = () => {
-  const {
-    routeData, 
-    setRouteData,
-    isInsideBuilding, 
-    setIsInsideBuilding,
-    myLocationString, 
-    setmyLocationString,
-    isTransit, 
-    setIsTransit,
-    highlightedBuilding, 
-    setHighlightedBuilding,
-    myLocationCoords, 
-    setMyLocationCoords
-  } = useCoords();
+// Mock the dependencies
+jest.mock('../src/data/CoordsContext');
+jest.mock('../src/data/IndoorContext');
+jest.mock('@rnmapbox/maps', () => ({
+  ShapeSource: jest.fn(() => null),
+  FillLayer: jest.fn(() => null),
+  SymbolLayer: jest.fn(() => null),
+  CircleLayer: jest.fn(() => null),
+}));
 
-  return (
-    <div>
-      <button data-testid="set-route-data" onClick={() => setRouteData({ test: 'route' })}>Set Route Data</button>
-      <button data-testid="set-inside-building" onClick={() => setIsInsideBuilding(true)}>Set Inside Building</button>
-      <button data-testid="set-location-string" onClick={() => setmyLocationString("Hall Building")}>Set Location String</button>
-      <button data-testid="set-is-transit" onClick={() => setIsTransit(true)}>Set Is Transit</button>
-      <button data-testid="set-highlighted-building" onClick={() => setHighlightedBuilding({ properties: { id: 'hall' } })}>Set Highlighted Building</button>
-      <button data-testid="set-location-coords" onClick={() => setMyLocationCoords({ latitude: 45.497, longitude: -73.578 })}>Set Location Coords</button>
-      
-      <div data-testid="route-data">{JSON.stringify(routeData)}</div>
-      <div data-testid="is-inside-building">{isInsideBuilding.toString()}</div>
-      <div data-testid="location-string">{myLocationString}</div>
-      <div data-testid="is-transit">{isTransit.toString()}</div>
-      <div data-testid="highlighted-building">{JSON.stringify(highlightedBuilding)}</div>
-      <div data-testid="location-coords">{JSON.stringify(myLocationCoords)}</div>
-    </div>
-  );
-};
+// Fix the import path to match the actual import path in HighlightIndoorMap
+jest.mock('../src/components/IndoorPointsOfInterest', () => ({
+  IndoorPointsOfInterest: jest.fn(() => null),
+}));
 
-// Test component for IndoorContext
-const TestIndoorComponent = () => {
-  const {
-    buildingHasFloors,
-    setBuildingHasFloors,
-    inFloorView,
-    setInFloorView,
-    currentFloor,
-    setCurrentFloor,
-    floorList,
-    setFloorList,
-    currentFloorAssociations,
-    setCurrentFloorAssociations,
-    indoorFeatures,
-    setIndoorFeatures
-  } = useIndoor();
-
-  return (
-    <div>
-      <button data-testid="set-building-has-floors" onClick={() => setBuildingHasFloors(true)}>Set Building Has Floors</button>
-      <button data-testid="set-in-floor-view" onClick={() => setInFloorView(true)}>Set In Floor View</button>
-      <button data-testid="set-current-floor" onClick={() => setCurrentFloor("1st Floor")}>Set Current Floor</button>
-      <button data-testid="set-floor-list" onClick={() => setFloorList(["1st Floor", "2nd Floor"])}>Set Floor List</button>
-      <button 
-        data-testid="set-floor-associations" 
-        onClick={() => setCurrentFloorAssociations([{ buildingID: 'hall', floor: '1', component: 'h1Features' }])}>
-        Set Floor Associations
-      </button>
-      <button 
-        data-testid="set-indoor-features" 
-        onClick={() => setIndoorFeatures([{ type: 'Feature', properties: { ref: 'H-101' } }])}>
-        Set Indoor Features
-      </button>
-      
-      <div data-testid="building-has-floors">{buildingHasFloors.toString()}</div>
-      <div data-testid="in-floor-view">{inFloorView.toString()}</div>
-      <div data-testid="current-floor">{currentFloor || 'null'}</div>
-      <div data-testid="floor-list">{JSON.stringify(floorList)}</div>
-      <div data-testid="floor-associations">{JSON.stringify(currentFloorAssociations)}</div>
-      <div data-testid="indoor-features">{JSON.stringify(indoorFeatures)}</div>
-    </div>
-  );
-};
-
-describe('CoordsContext', () => {
-  test('provides default values', () => {
-    render(
-      <CoordsProvider>
-        <TestCoordsComponent />
-      </CoordsProvider>
-    );
-    
-    expect(screen.getByTestId('route-data')).toHaveTextContent('null');
-    expect(screen.getByTestId('is-inside-building')).toHaveTextContent('false');
-    expect(screen.getByTestId('location-string')).toHaveTextContent('');
-    expect(screen.getByTestId('is-transit')).toHaveTextContent('false');
-    expect(screen.getByTestId('highlighted-building')).toHaveTextContent('null');
-    expect(screen.getByTestId('location-coords')).toHaveTextContent('null');
-  });
-
-  test('updates routeData state correctly', () => {
-    render(
-      <CoordsProvider>
-        <TestCoordsComponent />
-      </CoordsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-route-data').click();
-    });
-    
-    expect(screen.getByTestId('route-data')).toHaveTextContent('{"test":"route"}');
-  });
-
-  test('updates isInsideBuilding state correctly', () => {
-    render(
-      <CoordsProvider>
-        <TestCoordsComponent />
-      </CoordsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-inside-building').click();
-    });
-    
-    expect(screen.getByTestId('is-inside-building')).toHaveTextContent('true');
-  });
-
-  test('updates myLocationString state correctly', () => {
-    render(
-      <CoordsProvider>
-        <TestCoordsComponent />
-      </CoordsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-location-string').click();
-    });
-    
-    expect(screen.getByTestId('location-string')).toHaveTextContent('Hall Building');
-  });
-
-  test('updates isTransit state correctly', () => {
-    render(
-      <CoordsProvider>
-        <TestCoordsComponent />
-      </CoordsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-is-transit').click();
-    });
-    
-    expect(screen.getByTestId('is-transit')).toHaveTextContent('true');
-  });
-
-  test('updates highlightedBuilding state correctly', () => {
-    render(
-      <CoordsProvider>
-        <TestCoordsComponent />
-      </CoordsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-highlighted-building').click();
-    });
-    
-    expect(screen.getByTestId('highlighted-building')).toHaveTextContent('{"properties":{"id":"hall"}}');
-  });
-
-  test('updates myLocationCoords state correctly', () => {
-    render(
-      <CoordsProvider>
-        <TestCoordsComponent />
-      </CoordsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-location-coords').click();
-    });
-    
-    expect(screen.getByTestId('location-coords')).toHaveTextContent('{"latitude":45.497,"longitude":-73.578}');
-  });
-
-  test('provides updated context value when state changes', () => {
-    render(
-      <CoordsProvider>
-        <TestCoordsComponent />
-      </CoordsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-route-data').click();
-      screen.getByTestId('set-inside-building').click();
-      screen.getByTestId('set-location-string').click();
-      screen.getByTestId('set-is-transit').click();
-      screen.getByTestId('set-highlighted-building').click();
-      screen.getByTestId('set-location-coords').click();
-    });
-    
-    expect(screen.getByTestId('route-data')).toHaveTextContent('{"test":"route"}');
-    expect(screen.getByTestId('is-inside-building')).toHaveTextContent('true');
-    expect(screen.getByTestId('location-string')).toHaveTextContent('Hall Building');
-    expect(screen.getByTestId('is-transit')).toHaveTextContent('true');
-    expect(screen.getByTestId('highlighted-building')).toHaveTextContent('{"properties":{"id":"hall"}}');
-    expect(screen.getByTestId('location-coords')).toHaveTextContent('{"latitude":45.497,"longitude":-73.578}');
-  });
-});
-
-describe('IndoorContext', () => {
-  test('provides default values', () => {
-    render(
-      <IndoorsProvider>
-        <TestIndoorComponent />
-      </IndoorsProvider>
-    );
-    
-    expect(screen.getByTestId('building-has-floors')).toHaveTextContent('false');
-    expect(screen.getByTestId('in-floor-view')).toHaveTextContent('false');
-    expect(screen.getByTestId('current-floor')).toHaveTextContent('null');
-    expect(screen.getByTestId('floor-list')).toHaveTextContent('[]');
-    expect(screen.getByTestId('floor-associations')).toHaveTextContent('[]');
-    expect(screen.getByTestId('indoor-features')).toHaveTextContent('[]');
-  });
-
-  test('updates buildingHasFloors state correctly', () => {
-    render(
-      <IndoorsProvider>
-        <TestIndoorComponent />
-      </IndoorsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-building-has-floors').click();
-    });
-    
-    expect(screen.getByTestId('building-has-floors')).toHaveTextContent('true');
-  });
-
-  test('updates inFloorView state correctly', () => {
-    render(
-      <IndoorsProvider>
-        <TestIndoorComponent />
-      </IndoorsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-in-floor-view').click();
-    });
-    
-    expect(screen.getByTestId('in-floor-view')).toHaveTextContent('true');
-  });
-
-  test('updates currentFloor state correctly', () => {
-    render(
-      <IndoorsProvider>
-        <TestIndoorComponent />
-      </IndoorsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-current-floor').click();
-    });
-    
-    expect(screen.getByTestId('current-floor')).toHaveTextContent('1st Floor');
-  });
-
-  test('updates floorList state correctly', () => {
-    render(
-      <IndoorsProvider>
-        <TestIndoorComponent />
-      </IndoorsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-floor-list').click();
-    });
-    
-    expect(screen.getByTestId('floor-list')).toHaveTextContent('["1st Floor","2nd Floor"]');
-  });
-
-  test('updates currentFloorAssociations state correctly', () => {
-    render(
-      <IndoorsProvider>
-        <TestIndoorComponent />
-      </IndoorsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-floor-associations').click();
-    });
-    
-    expect(screen.getByTestId('floor-associations')).toHaveTextContent('[{"buildingID":"hall","floor":"1","component":"h1Features"}]');
-  });
-
-  test('updates indoorFeatures state correctly', () => {
-    render(
-      <IndoorsProvider>
-        <TestIndoorComponent />
-      </IndoorsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-indoor-features').click();
-    });
-    
-    expect(screen.getByTestId('indoor-features')).toHaveTextContent('[{"type":"Feature","properties":{"ref":"H-101"}}]');
-  });
-
-  test('provides updated context value when state changes', () => {
-    render(
-      <IndoorsProvider>
-        <TestIndoorComponent />
-      </IndoorsProvider>
-    );
-    
-    act(() => {
-      screen.getByTestId('set-building-has-floors').click();
-      screen.getByTestId('set-in-floor-view').click();
-      screen.getByTestId('set-current-floor').click();
-      screen.getByTestId('set-floor-list').click();
-      screen.getByTestId('set-floor-associations').click();
-      screen.getByTestId('set-indoor-features').click();
-    });
-    
-    expect(screen.getByTestId('building-has-floors')).toHaveTextContent('true');
-    expect(screen.getByTestId('in-floor-view')).toHaveTextContent('true');
-    expect(screen.getByTestId('current-floor')).toHaveTextContent('1st Floor');
-    expect(screen.getByTestId('floor-list')).toHaveTextContent('["1st Floor","2nd Floor"]');
-    expect(screen.getByTestId('floor-associations')).toHaveTextContent('[{"buildingID":"hall","floor":"1","component":"h1Features"}]');
-    expect(screen.getByTestId('indoor-features')).toHaveTextContent('[{"type":"Feature","properties":{"ref":"H-101"}}]');
-  });
-});
-
-// Test using both contexts together
-describe('Context Integration', () => {
-  const CombinedContextsComponent = () => {
-    const { highlightedBuilding, setHighlightedBuilding } = useCoords();
-    const { 
-      currentFloorAssociations, 
-      setCurrentFloorAssociations,
-      setBuildingHasFloors
-    } = useIndoor();
-    
-    // When highlightedBuilding changes, update currentFloorAssociations
-    React.useEffect(() => {
-      if (highlightedBuilding && highlightedBuilding.properties) {
-        const buildingID = highlightedBuilding.properties.id;
-        setCurrentFloorAssociations([
-          { buildingID, floor: '1', component: 'h1Features' }
-        ]);
-        setBuildingHasFloors(true);
-      }
-    }, [highlightedBuilding]);
-    
-    return (
-      <div>
-        <button 
-          data-testid="set-highlighted-building" 
-          onClick={() => setHighlightedBuilding({ properties: { id: 'hall' } })}>
-          Set Highlighted Building
-        </button>
-        <div data-testid="highlighted-building">{JSON.stringify(highlightedBuilding)}</div>
-        <div data-testid="floor-associations">{JSON.stringify(currentFloorAssociations)}</div>
-      </div>
-    );
-  };
+// This is the key change: properly mock the useIndoorFeatures hook
+jest.mock('../src/components/IndoorMap', () => {
+  const originalModule = jest.requireActual('../src/components/IndoorMap');
   
-  test('contexts work together properly', () => {
-    render(
-      <CoordsProvider>
-        <IndoorsProvider>
-          <CombinedContextsComponent />
-        </IndoorsProvider>
-      </CoordsProvider>
-    );
+  return {
+    ...originalModule,
+    useIndoorFeatures: jest.fn(),
+    HighlightIndoorMap: originalModule.HighlightIndoorMap
+  };
+});
+
+// Import the mocked useIndoorFeatures after it's been mocked
+import { useIndoorFeatures } from '../src/components/IndoorMap';
+
+// Create mock for floorNameFormat without mocking the entire module
+const floorNameFormat = (floor: string) => {
+  if (floor === '1') return '1st Floor';
+  if (floor === '2') return '2nd Floor';
+  return floor + (floor.includes('S') ? ' Floor' : 'th Floor');
+};
+
+describe('HighlightIndoorMap', () => {
+  // Setup common mock values
+  const mockSelectIndoorFeatures = jest.fn();
+  
+  beforeEach(() => {
+    jest.clearAllMocks();
     
-    // Check initial states
-    expect(screen.getByTestId('highlighted-building')).toHaveTextContent('null');
-    expect(screen.getByTestId('floor-associations')).toHaveTextContent('[]');
-    
-    // Trigger state change
-    act(() => {
-      screen.getByTestId('set-highlighted-building').click();
+    // Mock useCoords
+    (useCoords as jest.Mock).mockReturnValue({
+      highlightedBuilding: null,
+      isInsideBuilding: false,
+      destinationCoords: null,
+      myLocationCoords: null,
     });
     
-    // Check updated states reflecting integration between contexts
-    expect(screen.getByTestId('highlighted-building')).toHaveTextContent('{"properties":{"id":"hall"}}');
-    expect(screen.getByTestId('floor-associations')).toHaveTextContent('[{"buildingID":"hall","floor":"1","component":"h1Features"}]');
+    // Mock useIndoor
+    (useIndoor as jest.Mock).mockReturnValue({
+      setBuildingHasFloors: jest.fn(),
+      setInFloorView: jest.fn(),
+      inFloorView: false,
+      setCurrentFloor: jest.fn(),
+      setFloorList: jest.fn(),
+      currentFloorAssociations: [],
+      setCurrentFloorAssociations: jest.fn(),
+      setIndoorFeatures: jest.fn(),
+      indoorFeatures: [],
+      originRoom: null,
+      destinationRoom: null,
+      currentFloor: null,
+    });
+
+    // Here's the fixed mock implementation for useIndoorFeatures
+    (useIndoorFeatures as jest.Mock).mockReturnValue({
+      selectIndoorFeatures: mockSelectIndoorFeatures
+    });
+  });
+  
+  test('should not render any Mapbox components when indoorFeatures is empty', () => {
+    render(<HighlightIndoorMap />);
+    
+    expect(Mapbox.ShapeSource).not.toHaveBeenCalled();
+    expect(Mapbox.IndoorPointsOfInterest).not.toHaveBeenCalled();
+  });
+  
+  test('should render Mapbox components when indoorFeatures has data and inFloorView is true', () => {
+    // Setup mocks for this test
+    (useIndoor as jest.Mock).mockReturnValue({
+      setBuildingHasFloors: jest.fn(),
+      setInFloorView: jest.fn(),
+      inFloorView: true,
+      setCurrentFloor: jest.fn(),
+      setFloorList: jest.fn(),
+      currentFloorAssociations: [],
+      setCurrentFloorAssociations: jest.fn(),
+      setIndoorFeatures: jest.fn(),
+      indoorFeatures: [{ type: 'Feature', properties: {}, geometry: {} }],
+      originRoom: null,
+      destinationRoom: null,
+      currentFloor: null,
+    });
+    
+    render(<HighlightIndoorMap />);
+    
+    expect(Mapbox.ShapeSource).toHaveBeenCalledWith(
+      expect.objectContaining({
+        id: 'indoor-features',
+      }),
+      expect.anything()
+    );
+  });
+  
+  test('should render origin room pin when conditions met', () => {
+    // Setup mocks for this test
+    (useIndoor as jest.Mock).mockReturnValue({
+      setBuildingHasFloors: jest.fn(),
+      setInFloorView: jest.fn(),
+      inFloorView: true,
+      setCurrentFloor: jest.fn(),
+      setFloorList: jest.fn(),
+      currentFloorAssociations: [],
+      setCurrentFloorAssociations: jest.fn(),
+      setIndoorFeatures: jest.fn(),
+      indoorFeatures: [],
+      originRoom: {
+        ref: '115',
+        floor: '1',
+        coordinates: [-73.5789, 45.4966],
+      },
+      destinationRoom: null,
+      currentFloor: '1st Floor',
+    });
+    
+    render(<HighlightIndoorMap />);
+    
+    expect(Mapbox.ShapeSource).toHaveBeenCalledWith(
+      expect.objectContaining({
+        id: 'origin-room-pin',
+      }),
+      expect.anything()
+    );
+  });
+  
+  test('should render destination room pin when conditions met', () => {
+    // Setup mocks for this test
+    (useIndoor as jest.Mock).mockReturnValue({
+      setBuildingHasFloors: jest.fn(),
+      setInFloorView: jest.fn(),
+      inFloorView: true,
+      setCurrentFloor: jest.fn(),
+      setFloorList: jest.fn(),
+      currentFloorAssociations: [],
+      setCurrentFloorAssociations: jest.fn(),
+      setIndoorFeatures: jest.fn(),
+      indoorFeatures: [],
+      originRoom: null,
+      destinationRoom: {
+        ref: '835',
+        floor: '8',
+        coordinates: [-73.5789, 45.4966],
+      },
+      currentFloor: '8th Floor',
+    });
+    
+    render(<HighlightIndoorMap />);
+    
+    expect(Mapbox.ShapeSource).toHaveBeenCalledWith(
+      expect.objectContaining({
+        id: 'destination-room-pin',
+      }),
+      expect.anything()
+    );
+  });
+  
+  test('should not render room pins when floor does not match', () => {
+    // Setup mocks for this test
+    (useIndoor as jest.Mock).mockReturnValue({
+      setBuildingHasFloors: jest.fn(),
+      setInFloorView: jest.fn(),
+      inFloorView: true,
+      setCurrentFloor: jest.fn(),
+      setFloorList: jest.fn(),
+      currentFloorAssociations: [],
+      setCurrentFloorAssociations: jest.fn(),
+      setIndoorFeatures: jest.fn(),
+      indoorFeatures: [],
+      originRoom: {
+        ref: '115',
+        floor: '1',
+        coordinates: [-73.5789, 45.4966],
+      },
+      destinationRoom: {
+        ref: '835',
+        floor: '8',
+        coordinates: [-73.5789, 45.4966],
+      },
+      currentFloor: '9th Floor', // Different from both rooms
+    });
+    
+    render(<HighlightIndoorMap />);
+    
+    // Should not find a shape source with id containing 'pin'
+    const shapeSourceCalls = (Mapbox.ShapeSource as jest.Mock).mock.calls;
+    const hasPinCall = shapeSourceCalls.some((call: any) => 
+      call[0].id.includes('pin')
+    );
+    
+    expect(hasPinCall).toBe(false);
+  });
+  
+  test('should update floor associations when highlighted building changes', () => {
+    const setCurrentFloorAssociationsMock = jest.fn();
+    
+    (useCoords as jest.Mock).mockReturnValue({
+      highlightedBuilding: {
+        properties: {
+          id: 'H',
+        },
+      },
+      isInsideBuilding: false,
+      destinationCoords: null,
+      myLocationCoords: null,
+    });
+    
+    (useIndoor as jest.Mock).mockReturnValue({
+      setBuildingHasFloors: jest.fn(),
+      setInFloorView: jest.fn(),
+      inFloorView: false,
+      setCurrentFloor: jest.fn(),
+      setFloorList: jest.fn(),
+      currentFloorAssociations: [],
+      setCurrentFloorAssociations: setCurrentFloorAssociationsMock,
+      setIndoorFeatures: jest.fn(),
+      indoorFeatures: [],
+      originRoom: null,
+      destinationRoom: null,
+      currentFloor: null,
+    });
+    
+    render(<HighlightIndoorMap />);
+    
+    expect(setCurrentFloorAssociationsMock).toHaveBeenCalled();
+  });
+  
+  test('should set floor view when destination coords exist and inside building', () => {
+    const setInFloorViewMock = jest.fn();
+    
+    (useCoords as jest.Mock).mockReturnValue({
+      highlightedBuilding: null,
+      isInsideBuilding: true,
+      destinationCoords: [-73.5789, 45.4966],
+      myLocationCoords: [-73.5790, 45.4967],
+    });
+    
+    (useIndoor as jest.Mock).mockReturnValue({
+      setBuildingHasFloors: jest.fn(),
+      setInFloorView: setInFloorViewMock,
+      inFloorView: false,
+      setCurrentFloor: jest.fn(),
+      setFloorList: jest.fn(),
+      currentFloorAssociations: [],
+      setCurrentFloorAssociations: jest.fn(),
+      setIndoorFeatures: jest.fn(),
+      indoorFeatures: [],
+      originRoom: null,
+      destinationRoom: null,
+      currentFloor: null,
+    });
+    
+    render(<HighlightIndoorMap />);
+    
+    expect(setInFloorViewMock).toHaveBeenCalledWith(true);
+  });
+});
+
+// Testing the floorNameFormat function
+describe('floorNameFormat function', () => {
+  test('formats floor names correctly', () => {
+    expect(floorNameFormat('1')).toBe('1st Floor');
+    expect(floorNameFormat('2')).toBe('2nd Floor');
+    expect(floorNameFormat('3')).toBe('3th Floor');
+    expect(floorNameFormat('S2')).toBe('S2 Floor');
   });
 });
